% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{assoc_store_comp}
\alias{assoc_store_comp}
\title{Store/Competitor Associations}
\format{
A data frame with 1834 rows and 4 columns:
\describe{
  \item{assoc_key}{(integer) unique association id}
  \item{store_key}{(integer) unique store id}
  \item{comp_key}{(integer) unique competitor id}
  \item{is_primary}{(logical) indicates whether an association should be maintained with high priority}
  ...
}
}
\usage{
assoc_store_comp
}
\description{
A table of linkages between TWM and non-TWM locations, defining directed 'associations' between them.
}
\keyword{datasets}
