% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{banner_colors}
\alias{banner_colors}
\title{Banner Colors}
\format{
A data frame with 348 rows and 2 columns:
\describe{
  \item{banner}{(character) name of competitor's parent company}
  \item{color}{(character) RGB color code}
  ...
}
}
\usage{
banner_colors
}
\description{
A lookup table of RGB color values associated with non-TWM parent companies.
}
\keyword{datasets}
