% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{comp_location}
\alias{comp_location}
\title{Competitor Stores}
\format{
A data frame with 2037 rows and 8 columns:
\describe{
  \item{comp_key}{(integer) unique competitor id}
  \item{banner}{(character) name of competitor's parent company}
  \item{street_address}{(character) competitor location's street address}
  \item{city}{(character) competitor location's city}
  \item{state}{(character) competitor location's state}
  \item{zipcode}{(character) competitor location's zipcode}
  \item{latitude}{(numeric) latitude of competitor location}
  \item{longitude}{(numeric) longitude of competitor location}
  ...
}
}
\usage{
comp_location
}
\description{
A sample dataset of non-TWM physical locations and their attributes
}
\keyword{datasets}
