% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_compLoc.R
\name{compLocServer}
\alias{compLocServer}
\title{Server Function for Competitor Location Module}
\usage{
compLocServer(id)
}
\arguments{
\item{id}{module id}
}
\value{
shiny reactive
}
\description{
Server Function for Competitor Location Module
}
