% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_compLoc.R
\name{compLocUI}
\alias{compLocUI}
\title{UI for Competitor Location Module}
\usage{
compLocUI(id)
}
\arguments{
\item{id}{module id}
}
\value{
shiny.tag.list
}
\description{
UI for Competitor Location Module
}
