% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_locLeaflet.R
\name{locLeafletServer}
\alias{locLeafletServer}
\title{Server Function for Location Leaflet Module}
\usage{
locLeafletServer(id)
}
\arguments{
\item{id}{module id}
}
\value{
shiny reactive
}
\description{
Server Function for Location Leaflet Module
}
