% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_locLeaflet.R
\name{locLeafletUI}
\alias{locLeafletUI}
\title{UI for Location Leaflet Module}
\usage{
locLeafletUI(id)
}
\arguments{
\item{id}{module id}
}
\value{
shiny.tag.list
}
\description{
UI for Location Leaflet Module
}
