% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{store_location}
\alias{store_location}
\title{'Total Wine & More' Stores}
\format{
A data frame with 272 rows and 10 columns:
\describe{
  \item{store_key}{(integer) unique store id}
  \item{price_zone}{(character) name of store 'price zone'}
  \item{name}{(character) name of store}
  \item{description}{(character) long description of store}
  \item{street_address}{(character) store location's street address}
  \item{city}{(character) store location's city}
  \item{state}{(character) store location's state}
  \item{zipcode}{(character) store location's zipcode}
  \item{latitude}{(numeric) latitude of store location}
  \item{longitude}{(numeric) longitude of store location}
  ...
}
}
\usage{
store_location
}
\description{
A sample dataset of TWM physical locations and their attributes
}
\keyword{datasets}
