% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_storeLoc.R
\name{storeLocServer}
\alias{storeLocServer}
\title{Server Function for Store Location Module}
\usage{
storeLocServer(id)
}
\arguments{
\item{id}{module id}
}
\value{
shiny reactive
}
\description{
Server Function for Store Location Module
}
